+++ sql/core/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -71,6 +71,7 @@ object SQLConf {
       .createWithDefault(10)
 
   val ALLOW_MULTIPLE_CONTEXTS = SQLConfigBuilder("spark.sql.allowMultipleContexts")
+    .internal()
     .doc("When set to true, creating multiple SQLContexts/HiveContexts is allowed. " +
       "When set to false, only one SQLContext/HiveContext is allowed to be created " +
       "through the constructor (new SQLContexts/HiveContexts created through newSession " +
@@ -82,11 +83,12 @@ object SQLConf {
 
   val ALLOW_MULTIPLE_CONTEXTS_ERROR_MESSAGE =
     SQLConfigBuilder("spark.sql.allowMultipleContexts.errorMessage")
+      .internal()
       .doc("The error message used by the exception when spark.sql.allowMultipleContexts is " +
         "set to false.")
       .stringConf
       .createWithDefault(
-        "Only one SQLContext/HiveContext may be running in this JVM. " +
+        "Only one SparkSession/SQLContext/HiveContext may be running in this JVM. " +
         s"It is recommended to use SQLContext.getOrCreate to get the instantiated " +
         s"SQLContext/HiveContext. To ignore this error, " +
         s"set ${ALLOW_MULTIPLE_CONTEXTS.key} = true in SparkConf.")
