+++ sql/core/src/test/scala/org/apache/spark/sql/sources/DataSourceTest.scala
@@ -24,7 +24,7 @@ private[sql] abstract class DataSourceTest extends QueryTest {
 
   // We want to test some edge cases.
   protected lazy val caseInsensitiveContext: SQLContext = {
-    val ctx = new SQLContext(spark.sparkContext)
+    val ctx = new SQLContext(spark)
     ctx.setConf(SQLConf.CASE_SENSITIVE, false)
     ctx
   }
